refactor: Functions in [id].vue is put in a composable
